sdk: refactoring the GitHub OIDC stuff together
